add the begining of the network
