+++ apps/dragonfly_browser/lib/src/screens/developer_tools/developer_tools_screen.dart
@@ -1,4 +1,4 @@
-import 'package:dragonfly/src/screens/favorites/favorite_bar.dart';
+import 'package:dragonfly/src/screens/developer_tools/tools/network_tool_screen.dart';
 import 'package:dragonfly/src/screens/lobby/cubit/browser_interface_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -53,10 +53,7 @@ class _DeveloperToolsScreenState extends State<DeveloperToolsScreen> {
               ],
             ),
           ),
-          SizedBox(
-            height: 40,
-            child: ContentRequestSelectionBar(),
-          ),
+          Expanded(child: NetworkToolScreen()),
         ],
       ),
     );
@@ -92,94 +89,6 @@ class IconWithTool extends StatelessWidget {
           Icon(icon),
           Text(label),
         ],
-      ),
-    );
-  }
-}
-
-class ContentRequestSelectionBar extends StatefulWidget {
-  const ContentRequestSelectionBar({super.key});
-
-  @override
-  State<ContentRequestSelectionBar> createState() =>
-      _ContentRequestSelectionBarState();
-}
-
-class _ContentRequestSelectionBarState
-    extends State<ContentRequestSelectionBar> {
-  final contentRequests = ["All", "HTML", "CSS"];
-  List<bool> selectedContent = [true, false, false];
-
-  void _updateSelection(int index) {
-    setState(() {
-      if (index == 0) {
-        // If "All" is selected, deselect others
-        selectedContent = [true, false, false];
-      } else {
-        // If any other is selected, deselect "All" and toggle the clicked item
-        selectedContent[0] = false;
-        selectedContent[index] = !selectedContent[index];
-
-        // If all others are deselected, select "All"
-        if (!selectedContent.contains(true)) {
-          selectedContent[0] = true;
-        }
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 50,
-      padding: const EdgeInsets.symmetric(horizontal: 8.0),
-      child: ListView.builder(
-        itemCount: contentRequests.length,
-        scrollDirection: Axis.horizontal,
-        itemBuilder: (context, index) => GestureDetector(
-          onTap: () => _updateSelection(index),
-          child: ContentRequestSelection(
-            contentType: contentRequests[index],
-            isSelected: selectedContent[index],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class ContentRequestSelection extends StatelessWidget {
-  const ContentRequestSelection(
-      {super.key, required this.contentType, required this.isSelected});
-
-  final String contentType;
-  final bool isSelected;
-
-  @override
-  Widget build(BuildContext context) {
-    return MouseRegion(
-      cursor: SystemMouseCursors.click,
-      child: AnimatedContainer(
-        duration: Durations.short3,
-        padding: const EdgeInsets.symmetric(horizontal: 12.0, vertical: 8.0),
-        margin: const EdgeInsets.symmetric(horizontal: 4.0),
-        decoration: BoxDecoration(
-          border: Border(
-            bottom: BorderSide(
-              color: isSelected ? Colors.blue.shade700 : Colors.transparent,
-              width: 3,
-            ),
-          ),
-        ),
-        child: Center(
-          child: Text(
-            contentType,
-            style: TextStyle(
-              color: isSelected ? Colors.blue.shade700 : Colors.black87,
-              fontWeight: isSelected ? FontWeight.bold : FontWeight.normal,
-            ),
-          ),
-        ),
       ),
     );
   }
